backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
openai>=1.50
pydantic>=2.9
pyyaml>=6.0
slowapi>=0.1.9
python-frontmatter>=1.1
python-dotenv>=1.0
